feat: Added toast ARIA live region, conversation entry semantic roles,…
- src/webview/a11y.ts
- src/webview/index.ts
- src/webview/renderer.ts
- src/webview/message-handler.ts
- src/webview/message-handler.test.ts
- src/webview/__tests__/message-handler.test.ts

GSD-Task: S02/T01
